Rename rest_api::auth::roles to ..::rbac
This change renames the rest_api::auth::roles module to rbac, for
Role-based Access Control.  This more accurately reflects the purpose of
the module.
