+++ libsplinter/src/rest_api/auth/rbac/store/diesel/operations/add_role.rs
@@ -14,7 +14,7 @@
 
 use diesel::{dsl::insert_into, prelude::*};
 
-use crate::rest_api::auth::roles::store::{
+use crate::rest_api::auth::rbac::store::{
     diesel::{
         models::{RoleModel, RolePermissionModel},
         schema::{role_permissions, roles},
