+++ libsplinter/src/rest_api/auth/rbac/store/diesel/operations/update_role.rs
@@ -17,7 +17,7 @@ use diesel::{
     prelude::*,
 };
 
-use crate::rest_api::auth::roles::store::{
+use crate::rest_api::auth::rbac::store::{
     diesel::{
         models::{RoleModel, RolePermissionModel},
         schema::{role_permissions, roles},
