+++ libsplinter/src/rest_api/auth/rbac/store/error.rs
