+++ libsplinter/src/store/mod.rs
@@ -62,7 +62,7 @@ pub trait StoreFactory {
     #[cfg(feature = "authorization")]
     fn get_role_based_authorization_store(
         &self,
-    ) -> Box<dyn crate::rest_api::auth::roles::store::RoleBasedAuthorizationStore>;
+    ) -> Box<dyn crate::rest_api::auth::rbac::store::RoleBasedAuthorizationStore>;
 
     #[cfg(feature = "biome-profile")]
     fn get_biome_user_profile_store(&self) -> Box<dyn crate::biome::UserProfileStore>;
